app/auth.py
from datetime import datetime, timedelta, timezone
from typing import Optional, cast

from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from jose import JWTError, jwt
import bcrypt
from sqlalchemy.orm import Session

from app.database import get_db
from app.models import User

SECRET_KEY = "a_very_secret_key_for_this_lab_only"
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30

oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")

def verify_password(plain_password: str, hashed_password: str) -> bool:
    return bcrypt.checkpw(plain_password.encode('utf-8'), hashed_password.encode('utf-8'))

def get_password_hash(password: str) -> str:
    salt = bcrypt.gensalt()
    hashed = bcrypt.hashpw(password.encode('utf-8'), salt)
    return hashed.decode('utf-8')

# User authentication functions
def get_user_by_username(db: Session, username: str) -> Optional[User]:
    return db.query(User).filter(User.username == username).first()

def get_user_by_email(db: Session, email: str) -> Optional[User]:
    return db.query(User).filter(User.email == email).first()

def authenticate_user(db: Session, username: str, password: str) -> Optional[User]:
    user = get_user_by_username(db, username)
    if not user:
        return None
    # Cast SQLAlchemy Column to string for type checking
    hashed_password = cast(str, user.hashed_password)
    if not verify_password(password, hashed_password):
        return None
    return user

def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.now(timezone.utc) + expires_delta
    else:
        expire = datetime.now(timezone.utc) + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt

def decode_token(token: str) -> Optional[str]:
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        username: Optional[str] = payload.get("sub")
        if username is None:
            return None
        return username
    except JWTError:
        return None

async def get_current_user(token: str = Depends(oauth2_scheme), db: Session = Depends(get_db)) -> User:
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )

    username = decode_token(token)
    if username is None:
        raise credentials_exception

    user = get_user_by_username(db, username)
    if user is None:
        raise credentials_exception

    return user

async def get_user_from_token(token: str) -> str:
    """
    WebSocket authentication dependency - extracts username from JWT token.

    This function is specifically designed for WebSocket endpoints where the standard
    OAuth2PasswordBearer dependency cannot be used. WebSockets require a different
    authentication approach since they can't use standard HTTP headers.

    Args:
        token: JWT token string passed as query parameter

    Returns:
        str: Username extracted from the token

    Raises:
        HTTPException: If token is missing or invalid
    """
    # START Task 3.1
    # 1. Check if the token is None or empty. If so, raise an HTTPException with status_code 401 and detail "Authentication token is missing".
    if not token:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Authentication token is missing"
        )

    # 2. Use the decode_token(token) helper function to validate and decode the JWT.
    username = decode_token(token)

    # 3. If decode_token returns None, raise an HTTPException with status_code 401 and detail "Invalid authentication token".
    if username is None:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid authentication token"
        )

    # 4. If the token is valid, return the username.
    return username
    # END Task 3.1

async def get_websocket_user(token: str, db: Session) -> User:
    """Enhanced WebSocket authentication that validates user exists in database"""
    username = await get_user_from_token(token)
    user = get_user_by_username(db, username)
    if user is None:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="User not found"
        )
    return user
