pyproject.toml
[project]
name = "py-chat-websocket-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic==1.14.0",
    "anyio==4.10.0",
    "bcrypt>=5.0.0",
    "email-validator==2.3.0",
    "fastapi[standard]==0.116.2",
    "httpx==0.28.1",
    "pytest==8.4.2",
    "pytest-asyncio==1.2.0",
    "pytest-timeout==2.3.1",
    "python-jose==3.5.0",
    "python-multipart==0.0.20",
    "sqlalchemy==2.0.36",
    "starlette==0.48.0",
    "uvicorn==0.35.0",
    "websockets==15.0.1",
]
